Move createInitialState from Game to State
